BMI + GNSS work

+++ Header/Satellite.h
@@ -24,8 +24,8 @@
 //BMI085
 #define BMI_CS_A_GPIO			XMC_GPIO_PORT3 // 3.14	- 134
 #define BMI_CS_G_GPIO			XMC_GPIO_PORT4 // 4.3	- 121
-#define BMI_CS_A_PIN	(u8)	134
-#define BMI_CS_G_PIN	(u8)	121
+#define BMI_CS_A_PIN			134
+#define BMI_CS_G_PIN			121
 #define BMI_XMC_MODE 			XMC_GPIO_MODE_OUTPUT_PUSH_PULL
 
 //GNSS
